fix: fixing items overflow
