+++ pubspec.yaml
@@ -36,7 +36,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  feather_icons: ^1.2.0
   intl: ^0.17.0
 
 dev_dependencies:
